Module aws_sdk_servicecatalogappregistry::error

Modules

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

validation_exception

See ValidationException

Structs

AssociateAttributeGroupError

AssociateResourceError

ConflictException

There was a conflict when processing the request (for example, a resource with the given name already

CreateApplicationError

CreateAttributeGroupError

DeleteApplicationError

DeleteAttributeGroupError

DisassociateAttributeGroupError

DisassociateResourceError

GetApplicationError

GetAssociatedResourceError

GetAttributeGroupError

InternalServerException

The service is experiencing internal problems.

ListApplicationsError

ListAssociatedAttributeGroupsError

ListAssociatedResourcesError

ListAttributeGroupsError

ListTagsForResourceError

ResourceNotFoundException

The specified resource does not exist.

ServiceQuotaExceededException

The maximum number of resources per account has been reached.

SyncResourceError

TagResourceError

UntagResourceError

UpdateApplicationError

UpdateAttributeGroupError

ValidationException

The request has invalid or missing parameters.

Enums

AssociateAttributeGroupErrorKind

AssociateResourceErrorKind

CreateApplicationErrorKind

CreateAttributeGroupErrorKind

DeleteApplicationErrorKind

DeleteAttributeGroupErrorKind

DisassociateAttributeGroupErrorKind

DisassociateResourceErrorKind

GetApplicationErrorKind

GetAssociatedResourceErrorKind

GetAttributeGroupErrorKind

ListApplicationsErrorKind

ListAssociatedAttributeGroupsErrorKind

ListAssociatedResourcesErrorKind

ListAttributeGroupsErrorKind

ListTagsForResourceErrorKind

SyncResourceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateApplicationErrorKind

UpdateAttributeGroupErrorKind