Module aws_sdk_servicecatalogappregistry::model

Modules

application

See Application

application_summary

See ApplicationSummary

attribute_group

See AttributeGroup

attribute_group_summary

See AttributeGroupSummary

integrations

See Integrations

resource

See Resource

resource_group

See ResourceGroup

resource_info

See ResourceInfo

resource_integrations

See ResourceIntegrations

Structs

Application

Represents a Amazon Web Services Service Catalog AppRegistry application that is the top-level node in a hierarchy of related

ApplicationSummary

Summary of a Amazon Web Services Service Catalog AppRegistry application.

AttributeGroup

Represents a Amazon Web Services Service Catalog AppRegistry attribute group that is rich metadata which describes an application and its components.

AttributeGroupSummary

Summary of a Amazon Web Services Service Catalog AppRegistry attribute group.

Integrations

The information about the service integration.

Resource

The information about the resource.

ResourceGroup

The information about the resource group integration.

ResourceInfo

The information about the resource.

ResourceIntegrations

The service integration information about the resource.

Enums

ResourceGroupState

ResourceType

SyncAction