Module aws_sdk_servicecatalogappregistry::operation

Structs

AssociateAttributeGroup

Associates an attribute group with an application to augment the application's metadata

AssociateResource

Associates a resource with an application. Both the resource and the application can be specified either by ID or name.

CreateApplication

Creates a new application that is the top-level node in a hierarchy of related cloud resource abstractions.

CreateAttributeGroup

Creates a new attribute group as a container for user-defined attributes. This feature

DeleteApplication

Deletes an application that is specified either by its application ID or name. All associated attribute groups and resources must be disassociated from it before deleting an application.

DeleteAttributeGroup

Deletes an attribute group, specified either by its attribute group ID or name.

DisassociateAttributeGroup

Disassociates an attribute group from an application to remove the extra attributes contained in the attribute group from the application's metadata. This operation reverts AssociateAttributeGroup.

DisassociateResource

Disassociates a resource from application. Both the resource and the application can be specified either by ID or name.

GetApplication

Retrieves metadata information about one of your applications. The application can be specified either by its unique ID or by its name (which is unique within one account in one region at a given point in time). Specify by ID in automated workflows if you want to make sure that the exact same application is returned or a ResourceNotFoundException is thrown, avoiding the ABA addressing problem.

GetAssociatedResource

Gets the resource associated with the application.

GetAttributeGroup

Retrieves an attribute group, either by its name or its ID. The attribute group can be specified either by its unique ID or by its name.

ListApplications

Retrieves a list of all of your applications. Results are paginated.

ListAssociatedAttributeGroups

Lists all attribute groups that are associated with specified application. Results are paginated.

ListAssociatedResources

Lists all resources that are associated with specified application. Results are paginated.

ListAttributeGroups

Lists all attribute groups which you have access to. Results are paginated.

ListTagsForResource

Lists all of the tags on the resource.

SyncResource

Syncs the resource with current AppRegistry records.

TagResource

Assigns one or more tags (key-value pairs) to the specified resource.

UntagResource

Removes tags from a resource.

UpdateApplication

Updates an existing application with new attributes.

UpdateAttributeGroup

Updates an existing attribute group with new details.